+++ pkg/api/handler_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/kyverno/policy-reporter-kyverno-plugin/pkg/kyverno"
 )
 
-func Test_PolicyReportAPI(t *testing.T) {
+func Test_PolicyAPI(t *testing.T) {
 	t.Run("Empty Respose", func(t *testing.T) {
 		req, err := http.NewRequest("GET", "/policies", nil)
 		if err != nil {
@@ -38,16 +38,16 @@ func Test_PolicyReportAPI(t *testing.T) {
 			t.Fatal(err)
 		}
 
-		result := kyverno.Rule{
+		result := &kyverno.Rule{
 			ValidateMessage: "validation error: requests and limits required. Rule autogen-check-for-requests-and-limits failed at path /spec/template/spec/containers/0/resources/requests/",
 			Name:            "autogen-check-for-requests-and-limits",
 		}
 
-		policy := kyverno.Policy{
+		policy := &kyverno.Policy{
 			Kind:              "Policy",
 			Name:              "require-ressources",
 			Namespace:         "test",
-			Rules:             []kyverno.Rule{result},
+			Rules:             []*kyverno.Rule{result},
 			CreationTimestamp: time.Now(),
 		}
 
@@ -71,38 +71,14 @@ func Test_PolicyReportAPI(t *testing.T) {
 }
 
 func Test_HealthzAPI(t *testing.T) {
-	t.Run("Unhealthy Respose", func(t *testing.T) {
-		req, err := http.NewRequest("GET", "/healthz", nil)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		healthy := false
-
-		rr := httptest.NewRecorder()
-		handler := http.HandlerFunc(api.HealthzHandler(&healthy))
-
-		handler.ServeHTTP(rr, req)
-
-		if status := rr.Code; status != http.StatusServiceUnavailable {
-			t.Errorf("handler returned wrong status code: got %v want %v", status, http.StatusOK)
-		}
-
-		expected := `{ "error": "Service unhealthy" }`
-		if rr.Body.String() != expected {
-			t.Errorf("handler returned unexpected body: got %v want %v", rr.Body.String(), expected)
-		}
-	})
 	t.Run("Success Respose", func(t *testing.T) {
 		req, err := http.NewRequest("GET", "/healthz", nil)
 		if err != nil {
 			t.Fatal(err)
 		}
 
-		healthy := true
-
 		rr := httptest.NewRecorder()
-		handler := http.HandlerFunc(api.HealthzHandler(&healthy))
+		handler := http.HandlerFunc(api.HealthzHandler(map[string]bool{"Policy": true}))
 
 		handler.ServeHTTP(rr, req)
 
@@ -113,6 +89,22 @@ func Test_HealthzAPI(t *testing.T) {
 		expected := `{}`
 		if !strings.Contains(rr.Body.String(), expected) {
 			t.Errorf("handler returned unexpected body: got %v want %v", rr.Body.String(), expected)
+		}
+	})
+
+	t.Run("Unavailable Respose", func(t *testing.T) {
+		req, err := http.NewRequest("GET", "/healthz", nil)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		rr := httptest.NewRecorder()
+		handler := api.HealthzHandler(map[string]bool{})
+
+		handler.ServeHTTP(rr, req)
+
+		if status := rr.Code; status != http.StatusServiceUnavailable {
+			t.Errorf("handler returned wrong status code: got %v want %v", status, http.StatusServiceUnavailable)
 		}
 	})
 }