+++ pkg/kyverno/model.go
@@ -4,6 +4,23 @@ import (
 	"time"
 )
 
+// Event Enum
+type Event = int
+
+// Possible Policy Event Enums
+const (
+	Added Event = iota
+	Updated
+	Deleted
+)
+
+// LifecycleEvent of Policys
+type LifecycleEvent struct {
+	Type      Event
+	NewPolicy *Policy
+	OldPolicy *Policy
+}
+
 // Rule from the Policy spec clusterpolicies.kyverno.io/v1.Policy
 type Rule struct {
 	ValidateMessage string `json:"message,omitempty"`
@@ -19,16 +36,11 @@ type Policy struct {
 	AutogenControllers      []string  `json:"autogenControllers,omitempty"`
 	ValidationFailureAction string    `json:"validationFailureAction,omitempty"`
 	Background              bool      `json:"background"`
-	Rules                   []Rule    `json:"rules"`
+	Rules                   []*Rule   `json:"rules"`
 	Category                string    `json:"category,omitempty"`
 	Description             string    `json:"description,omitempty"`
 	Severity                string    `json:"severity,omitempty"`
 	CreationTimestamp       time.Time `json:"creationTimestamp,omitempty"`
 	UID                     string    `json:"uid,omitempty"`
 	Content                 string    `json:"content"`
-}
-
-// GetIdentifier returns a global unique Policy identifier
-func (pr Policy) GetIdentifier() string {
-	return pr.UID
 }