+++ pkg/listener/metrics.go
@@ -1,4 +1,4 @@
-package metrics
+package listener
 
 import (
 	"strconv"
@@ -6,41 +6,40 @@ import (
 	"github.com/kyverno/policy-reporter-kyverno-plugin/pkg/kyverno"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
-	"k8s.io/apimachinery/pkg/watch"
 )
 
-// CreatePolicyMetricsCallback for PolicyReport watch.Events
-func CreatePolicyMetricsCallback() kyverno.PolicyCallback {
+// CreatePolicyMetricsCallback for Policy watch.Events
+func NewPolicyMetricsListener() kyverno.PolicyListener {
 	policyGauge := promauto.NewGaugeVec(prometheus.GaugeOpts{
-		Name: "policy_report_kyverno_policy",
+		Name: "kyverno_policy",
 		Help: "List of all Policies",
 	}, []string{"namespace", "kind", "policy", "rule", "type", "background", "severity", "category", "validationFailureAction"})
 
 	prometheus.Register(policyGauge)
 
-	return func(event watch.EventType, policy kyverno.Policy, oldPolicy kyverno.Policy) {
-		switch event {
-		case watch.Added:
-			for _, rule := range policy.Rules {
-				policyGauge.With(generateResultLabels(policy, rule)).Set(1)
+	return func(event kyverno.LifecycleEvent) {
+		switch event.Type {
+		case kyverno.Added:
+			for _, rule := range event.NewPolicy.Rules {
+				policyGauge.With(generateResultLabels(event.NewPolicy, rule)).Set(1)
 			}
-		case watch.Modified:
-			for _, rule := range oldPolicy.Rules {
-				policyGauge.Delete(generateResultLabels(oldPolicy, rule))
+		case kyverno.Updated:
+			for _, rule := range event.OldPolicy.Rules {
+				policyGauge.Delete(generateResultLabels(event.OldPolicy, rule))
 			}
 
-			for _, rule := range policy.Rules {
-				policyGauge.With(generateResultLabels(policy, rule)).Set(1)
+			for _, rule := range event.NewPolicy.Rules {
+				policyGauge.With(generateResultLabels(event.NewPolicy, rule)).Set(1)
 			}
-		case watch.Deleted:
-			for _, rule := range policy.Rules {
-				policyGauge.Delete(generateResultLabels(policy, rule))
+		case kyverno.Deleted:
+			for _, rule := range event.NewPolicy.Rules {
+				policyGauge.Delete(generateResultLabels(event.NewPolicy, rule))
 			}
 		}
 	}
 }
 
-func generateResultLabels(policy kyverno.Policy, rule kyverno.Rule) prometheus.Labels {
+func generateResultLabels(policy *kyverno.Policy, rule *kyverno.Rule) prometheus.Labels {
 	labels := prometheus.Labels{
 		"namespace":               policy.Namespace,
 		"kind":                    policy.Kind,