+++ pkg/listener/metrics_test.go
@@ -1,82 +1,32 @@
-package metrics_test
+package listener_test
 
 import (
 	"fmt"
 	"strconv"
 	"testing"
 
-	"github.com/kyverno/policy-reporter-kyverno-plugin/pkg/kubernetes"
 	"github.com/kyverno/policy-reporter-kyverno-plugin/pkg/kyverno"
-	"github.com/kyverno/policy-reporter-kyverno-plugin/pkg/metrics"
+	"github.com/kyverno/policy-reporter-kyverno-plugin/pkg/listener"
 	"github.com/prometheus/client_golang/prometheus"
 	io_prometheus_client "github.com/prometheus/client_model/go"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime/serializer/yaml"
-	"k8s.io/apimachinery/pkg/watch"
 )
-
-var policy = `
-apiVersion: kyverno.io/v1
-kind: ClusterPolicy
-metadata:
-  annotations:
-    pod-policies.kyverno.io/autogen-controllers: Deployment
-    meta.helm.sh/release-name: kyverno
-    meta.helm.sh/release-namespace: kyverno
-    policies.kyverno.io/severity: medium
-    policies.kyverno.io/category: Pod Security Standards (Default)
-    policies.kyverno.io/description: HostPath volumes let pods use host directories
-      and volumes in containers. Using host resources can be used to access shared
-      data or escalate privileges and should not be allowed.
-  creationTimestamp: "2021-03-31T13:42:01Z"
-  generation: 1
-  labels:
-    app.kubernetes.io/managed-by: Helm
-  name: disallow-host-path
-  resourceVersion: "61655872"
-  uid: 953b1167-1ff5-4cf6-b636-3b7d0c0dd6c7
-spec:
-  background: true
-  rules:
-  - match:
-      resources:
-        kinds:
-        - Pod
-    name: host-path
-    validate:
-      message: HostPath volumes are forbidden. The fields spec.volumes[*].hostPath
-        must not be set.
-      pattern:
-        spec:
-          =(volumes):
-          - X(hostPath): "null"
-  validationFailureAction: audit
-`
-
-func NewPolicy() kyverno.Policy {
-	obj := unstructured.Unstructured{}
-	dec := yaml.NewDecodingSerializer(unstructured.UnstructuredJSONScheme)
-	dec.Decode([]byte(policy), nil, &obj)
-
-	return kubernetes.NewMapper().MapPolicy(obj.Object)
-}
 
 func Test_PolicyMetricGeneration(t *testing.T) {
 	pol1 := NewPolicy()
 	pol2 := NewPolicy()
 	pol2.Name = pol2.Name + "-updated"
 
-	handler := metrics.CreatePolicyMetricsCallback()
+	handler := listener.NewPolicyMetricsListener()
 
 	t.Run("Added Metric", func(t *testing.T) {
-		handler(watch.Added, pol1, kyverno.Policy{})
+		handler(kyverno.LifecycleEvent{Type: kyverno.Added, NewPolicy: pol1, OldPolicy: &kyverno.Policy{}})
 
 		metricFam, err := prometheus.DefaultGatherer.Gather()
 		if err != nil {
 			t.Errorf("Unexpected Error: %s", err)
 		}
 
-		results := findMetric(metricFam, "policy_report_kyverno_policy")
+		results := findMetric(metricFam, "kyverno_policy")
 
 		metricResult := results.GetMetric()
 		if err = testResultMetricLabels(metricResult[0], pol1); err != nil {
@@ -85,15 +35,15 @@ func Test_PolicyMetricGeneration(t *testing.T) {
 	})
 
 	t.Run("Modified Metric", func(t *testing.T) {
-		handler(watch.Added, pol1, kyverno.Policy{})
-		handler(watch.Modified, pol2, pol1)
+		handler(kyverno.LifecycleEvent{Type: kyverno.Added, NewPolicy: pol1, OldPolicy: &kyverno.Policy{}})
+		handler(kyverno.LifecycleEvent{Type: kyverno.Updated, NewPolicy: pol2, OldPolicy: pol1})
 
 		metricFam, err := prometheus.DefaultGatherer.Gather()
 		if err != nil {
 			t.Errorf("Unexpected Error: %s", err)
 		}
 
-		results := findMetric(metricFam, "policy_report_kyverno_policy")
+		results := findMetric(metricFam, "kyverno_policy")
 
 		metricResult := results.GetMetric()
 		if err = testResultMetricLabels(metricResult[0], pol2); err != nil {
@@ -102,23 +52,23 @@ func Test_PolicyMetricGeneration(t *testing.T) {
 	})
 
 	t.Run("Deleted Metric", func(t *testing.T) {
-		handler(watch.Added, pol1, kyverno.Policy{})
-		handler(watch.Modified, pol2, pol1)
-		handler(watch.Deleted, pol2, pol2)
+		handler(kyverno.LifecycleEvent{Type: kyverno.Added, NewPolicy: pol1, OldPolicy: &kyverno.Policy{}})
+		handler(kyverno.LifecycleEvent{Type: kyverno.Updated, NewPolicy: pol2, OldPolicy: pol1})
+		handler(kyverno.LifecycleEvent{Type: kyverno.Deleted, NewPolicy: pol2, OldPolicy: &kyverno.Policy{}})
 
 		metricFam, err := prometheus.DefaultGatherer.Gather()
 		if err != nil {
 			t.Errorf("Unexpected Error: %s", err)
 		}
 
-		results := findMetric(metricFam, "policy_report_kyverno_policy")
+		results := findMetric(metricFam, "kyverno_policy")
 		if results != nil {
 			t.Error("policy_report_kyverno_policy shoud no longer exist", *results.Name)
 		}
 	})
 }
 
-func testResultMetricLabels(metric *io_prometheus_client.Metric, policy kyverno.Policy) error {
+func testResultMetricLabels(metric *io_prometheus_client.Metric, policy *kyverno.Policy) error {
 	rule := policy.Rules[0]
 
 	if name := *metric.Label[0].Name; name != "background" {