+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.0.0
+
+* Use one HTTP Server with Port `8080` for Metrics and REST API
+* Use informer instead of watch
+* Improve Pub/Sub pattern
+* Improve Healthz API
+
 ## 0.1.0
 
 * Prometheus Metrics for active Policies